refactor: convert patterns to regular classes
The equals/hashCode must be the ID. Otherwise, multiple exact same patterns
would only occur once in the network pattern set.
And when you remove a pattern at that point,
even if there is another still present,
the pattern would be removed entirely from the
set since it's stored once because it's
the same pattern, even if duplicated.

This also makes it easier to cache the
output resources properly.