refactor: remove zeroed system for grid resources
It was necessary because the backing list
can never contain zero of a resource,
and because the grid resources used to keep
a reference to the resource amount in the list.
So we needed the zeroed system to properly show
"0" in the gui. However, now we just query the
backing list for the amount, which can be 0.