chore: upgrade refined architect
To fix issues with NeoForge asset
updating.

With this change, every api module is now
compiled with the mod instead of using
the fabric jar-in-jar system.
The reason for this is for consistency
with neoforge, which always had the apis
compiled along with the main mod.

The only reason why we used jij on fabric
was for legacy reasons
 (RS 2 was originally made on fabric)
 and for inclusion
of the apis in modmenu.
But this is not important.