feat: storage monitor autocrafting
Only if the resource could no longer be extracted,
try trigger autocrafting.
I tried to reuse shift clicking for this
to force autocrafting,
but shift is already used to only extract
one of a resource.

You have to wait 250ms until
the last extract to trigger autocrafting
as attack method from block might be called
very often, and we don't want open
the autocrafting window when it's not needed
(eg right after an extract).